Added the plot for timing

+++ Driver.h
@@ -10,10 +10,10 @@ class Driver {
         /* 
          * Perform multiple additions and returns the time taken
          */ 
-        double timeAdditions(int n);
-        double timeAdditions(int n, int maxN);
-        double timeDeletions(int n);
-        double timeLookup(int n);
+        long long timeAdditions(int n);
+        long long timeAdditions(int n, int maxN);
+        long long timeDeletions(int n);
+        long long timeLookup(int n);
 
 
 };
